requirements.txt
python-dotenv<=0.15
Flask==2.0.2
flask_socketio==5.3.6